po: rebuild the .gmo files in the debian/rules file
This commit reverts commit aff534958b0 and transfers the
responsibility of rebuilding the .gmo files to the debian/rules file,
as updating the gmo files can cause problems for other distribution's
building systems.  (We do this because we don't update the .gmo files
when doing a mid-release snapshot, since binary files can't be
represented in diff files.  So we only update the .po files, and let
the package build process take care of the .gmo files.)

Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>