libext2fs: allocate separate memory regions for each inode in the cache
The changes to support metadata checksum allocated a single large
array for all of the inodes in the inode cache.  This is slightly more
efficient, but given that the inode cache is small (only 4 inodes) it
doesn't really have that much benefit.  The problem with doing things
this way is that the memory overruns, such as the one fixed in commit
43c4910371a, do not get detected by valgrind.

Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>