libext2fs: add space for checksum when unconverting a hashed directory block
The ext2fs_link function has the unfortunate habit of converting
hashed directories into unhashed directories.  It doesn't notice that
it's slicing and dicing directory entries from a former dx_{root,node}
block, and therefore doesn't write a protective dirent into the end of
the block to store the checksum.  Teach it to do this.

Signed-off-by: Darrick J. Wong <darrick.wong@oracle.com>
Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>