libext2fs: handle inline_data in block iterator by returning an error code
An inode with inline data has no data blocks, so we can not iterate
over such an inode.  Return an error code which indicates this fact;
callers can use this to determine whether or not the inode has inline
data, and then call some routine to iterate over the directory intries
in the line data or read the inline data, as appropriate.

Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>